temporary move stats page routes
